pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "bcrypt>=4.3.0",
    "flask>=3.1.0",
    "gitpython>=3.1.44",
    "pyjwt>=2.10.1",
    "pymongo>=4.11.2",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "twilio>=9.4.6",
    "websockets>=15.0.1",
]
